Changed partyid to bigint

+++ src/Storage/Migration/v0.25/01-create-tables.sql
@@ -2,7 +2,7 @@ CREATE TABLE IF NOT EXISTS storage.outbox
 (
 	instanceid UUID PRIMARY KEY,
 	appid TEXT NOT NULL,
-	partyid INTEGER NOT NULL,
+	partyid BIGINT NOT NULL,
     instancecreated TIMESTAMPTZ NOT NULL,
     validfrom TIMESTAMPTZ NOT NULL,
     ismigration BOOLEAN NOT NULL DEFAULT FALSE,
